X64: Changed 0x%x formats in log.cc to 0x%p and omitted reinterpretting pointers.
Review URL: http://codereview.chromium.org/114010


git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@1893 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/log.h
@@ -261,12 +261,12 @@ class Logger {
 // Class that extracts stack trace, used for profiling.
 class StackTracer BASE_EMBEDDED {
  public:
-  explicit StackTracer(unsigned int low_stack_bound)
+  explicit StackTracer(uintptr_t low_stack_bound)
       : low_stack_bound_(low_stack_bound) { }
   void Trace(TickSample* sample);
  private:
 
-  unsigned int low_stack_bound_;
+  uintptr_t low_stack_bound_;
 };
